Make it compile with LLVM
Hello,
Nowadays the official compiler on FreeBSD is LLVM; LLVM on FreeBSD is a bit picky about standard conformance: if you say "-D_POSIX_C_SOURCE=999999L" it means to compile making visible ONLY what is available in "the latest version on POSIX". In example "isascii()" is currecntly deprecated.
The PR should not affect any other platform, but I could not test on Linux.
With this change and the PR I submitted a few hours ago for tkrzw-rpc both things compile happily on FreeBSD/LLVM, link with systems-provided gRPC, abseil and protobuf ports, and so far seem to be quite stable.