Move static library dependencies to pkg-config Libs.private
An application only needs to directly link with tkrzw's dependencies when
linking statically.  Don't list those when pkg-config --libs is invoked
without --static.

Fixes link failures when installed from RPM on Fedora.  The tkrzw-devel
package doesn't depend on lz4-devel (and wouldn't ordinarily need to) so
the liblz4.so symlink may not be present.